requirements.txt
Django
djangorestframework
streamlit
groq
python-dotenv
requests
PyPDF2
python-docx
pdfplumber
django-cors-headers

